// schemas/auth.ts
import { z } from 'zod';

export const loginSchema = z.object({
  email: z.email('Informe um e-mail válido'),
  password: z.string().min(6, 'Mínimo de 6 caracteres'),
});

export const registerSchema = z.object({
  name: z.string().min(2, 'Informe seu nome'),
  email: z.email('Informe um e-mail válido'),
  password: z.string().min(6, 'Mínimo de 6 caracteres'),
});

export const forgotPasswordSchema = z.object({
  email: z.email('Informe um e-mail válido'),
});

export type LoginInput = z.infer<typeof loginSchema>;
export type RegisterInput = z.infer<typeof registerSchema>;